+++ mobile/lib/ui/components/home_header_widget.dart
@@ -10,8 +10,6 @@ import 'package:photos/ui/components/buttons/icon_button_widget.dart';
 import "package:photos/ui/settings/backup/backup_folder_selection_page.dart";
 import "package:photos/utils/dialog_util.dart";
 import "package:photos/utils/navigation_util.dart";
-import "package:photos/utils/photo_manager_util.dart";
-
 class HomeHeaderWidget extends StatefulWidget {
   final Widget centerWidget;
   const HomeHeaderWidget({required this.centerWidget, super.key});
@@ -48,7 +46,7 @@ class _HomeHeaderWidgetState extends State<HomeHeaderWidget> {
           onTap: () async {
             try {
               final PermissionState state =
-                  await requestPhotoMangerPermissions();
+                  await permissionService.requestPhotoMangerPermissions();
               await permissionService.onUpdatePermission(state);
             } on Exception catch (e) {
               Logger("HomeHeaderWidget").severe(
