+++ mobile/lib/ui/home/start_backup_hook_widget.dart
@@ -3,7 +3,7 @@ import "dart:async";
 import 'package:flutter/material.dart';
 import 'package:photo_manager/photo_manager.dart';
 import 'package:photos/generated/l10n.dart';
-import 'package:photos/services/sync/local_sync_service.dart';
+import "package:photos/service_locator.dart";
 import 'package:photos/ui/common/gradient_button.dart';
 import 'package:photos/ui/settings/backup/backup_folder_selection_page.dart';
 import 'package:photos/utils/navigation_util.dart';
@@ -42,8 +42,7 @@ class StartBackupHookWidget extends StatelessWidget {
               padding: const EdgeInsets.fromLTRB(20, 0, 20, 0),
               child: GradientButton(
                 onTap: () async {
-                  if (LocalSyncService.instance
-                      .hasGrantedLimitedPermissions()) {
+                  if (permissionService.hasGrantedLimitedPermissions()) {
                     unawaited(PhotoManager.presentLimited());
                   } else {
                     // ignore: unawaited_futures
