--- mobile/lib/utils/photo_manager_util.dart
@@ -1,12 +1,0 @@
-import "package:photo_manager/photo_manager.dart";
-
-Future<PermissionState> requestPhotoMangerPermissions() {
-  return PhotoManager.requestPermissionExtend(
-    requestOption: const PermissionRequestOption(
-      androidPermission: AndroidPermission(
-        type: RequestType.common,
-        mediaLocation: true,
-      ),
-    ),
-  );
-}
